Cleanup and refactor tests
