+++ core-s3/src/test/scala/io/aiven/guardian/kafka/s3/MinioS3Test.scala
@@ -1,0 +1,26 @@
+package io.aiven.guardian.kafka.s3
+
+import akka.stream.alpakka.s3.{AccessStyle, S3Settings}
+import akka.testkit.TestKitBase
+import com.dimafeng.testcontainers.ForAllTestContainer
+import org.scalatest.Suite
+import software.amazon.awssdk.auth.credentials.{AwsBasicCredentials, StaticCredentialsProvider}
+import software.amazon.awssdk.regions.Region
+import software.amazon.awssdk.regions.providers.AwsRegionProvider
+
+trait MinioS3Test extends ForAllTestContainer with TestKitBase { this: Suite =>
+  val S3DummyAccessKey = "DUMMY_ACCESS_KEY"
+  val S3DummySecretKey = "DUMMY_SECRET_KEY"
+
+  lazy val s3Settings = S3Settings()
+    .withEndpointUrl(s"http://${container.getHostAddress}")
+    .withCredentialsProvider(
+      StaticCredentialsProvider.create(AwsBasicCredentials.create(S3DummyAccessKey, S3DummySecretKey))
+    )
+    .withS3RegionProvider(new AwsRegionProvider {
+      lazy val getRegion: Region = Region.US_EAST_1
+    })
+    .withAccessStyle(AccessStyle.PathAccessStyle)
+
+  override val container: MinioContainer = new MinioContainer(S3DummyAccessKey, S3DummySecretKey)
+}
